requirements.txt
acme==1.22.0
amqp==5.1.1
anyjson==0.3.3
asgiref==3.4.1
asyncio==3.4.3
billiard==3.6.4.0
celery==5.2.6
certbot==1.22.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
ConfigArgParse==1.5.3
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Deprecated==1.2.13
distro==1.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.9
django-celery-beat==2.2.1
django-cors-headers==3.10.1
django-elasticsearch-dsl==7.1.2
django-elasticsearch-dsl-drf==0.22.4
django-extensions==3.1.5
django-filter==21.1
django-haystack==2.4.1
django-model-utils==4.2.0
django-modeltranslation==0.17.5
django-nested-inline==0.4.4
django-nine==0.2.5
django-parler-rest==2.1
django-redis==5.2.0
django-smart-selects==1.5.9
django-timezone-field==4.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
elasticsearch==7.10.1
elasticsearch-dsl==7.4.0
et-xmlfile==1.1.0
flower==1.0.0
graphviz==0.19.1
humanize==4.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
josepy==1.11.0
kombu==5.2.4
MarkupSafe==2.0.1
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
parsedatetime==2.6
Pillow==9.1.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pydot==1.4.2
pyflakes==2.4.0
PyJWT==2.3.0
pymongo==4.0.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyRFC3339==1.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
tornado==6.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.14.1
zope.component==5.0.1
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
